Fraud Detection/requirements.txt
kafka-python
pyspark
psycopg2
apache-airflow
docker
terraform
pandas